requirements.txt
streamlit
ebooklib
beautifulsoup4
PyMuPDF
pytesseract
Pillow